Refactor LGPO related tests
