+++ tests/pytests/unit/modules/win_lgpo/test_secedit_policy.py
